Metadata rework, part 1
The metadata for WOZ and 2IMG was exposed as a mix of properties and
name/value dictionary access.  While the disk image classes were
responsible for validating the inputs, the application still had a
lot of what-goes-where knowledge.

This defines a general-purpose metadata interface that can be used
to query objects with metadata.  The WOZ class has been converted,
and the sm/gm commands have been updated (temporarily breaking 2IMG
metadata access).