Implement app settings
Load and save the settings file, which lives in the same directory
as the executables (unless you're in a development environment, in
which case it lives in the project root).

We currently save and restore the main window size and position,
the internal window sizes and states, and the auto-open depth.  The
work file open code now follows the same depth rules as the cp2
"catalog" command.