Expand tree when items double-clicked
The archive tree in the top-left window is populated when the archive
is first opened, based on the "auto-open depth" setting.  If the
setting is below max, or if a disk image or file archive is added
after the file is opened, then some things may not be represented.

Now, double-clicking on a disk image or file archive will open that
entry as such, and add a new item to the file structure TreeView.
Similarly, double-clicking on an entry in the partition map shown
for multi-part formats will open that partition.  If these things
are already open, the focus will jump there.

This allows a full traversal from depth=shallow, except for embedded
or hybrid volumes, which don't show up unless explicitly searched for.
We'll need a menu item for that.