Add "scan for sub-volumes" action
This is only useful when auto open depth is "shallow".  It scans for
hybrid and embedded volumes and adds them to the tree according to
the current depth rule.  If the multi-part layout is already in the
tree, we move the selection there.

This required adding right-click actions to the archive tree window.
A non-functional "edit blocks/sectors" was also added.

Also, added a bit of documentation on cp2_wpf.  Made some updates to
various bits of documentation.