Improve info panel contents
The IChunkAccess available from the IDiskImage or Partition is now
marked read-only rather than unavailable, to make it easier to query
stats.  The underlying chunk object is passed to the IFileSystem and
published as RawAccess, so we're not actually making something new
available.

(It's in two places because you can have a filesystem instance that
is not part of a disk image, and disk images that don't have known
filesystems.)