Enable writing in sector editor
This is a little tricky because we don't want to allow editing of an
active filesystem, so before we modify a disk we need to close the
associated multi-part layout or filesystem, which requires closing
every disk image and file archive on them that are currently open and
visible in the archive tree.  This requires truncating the various
trees when "enable writes" is clicked, and restoring them when the
sector-edit dialog is closed.

Fully restoring the previous state doesn't seem necessary, so we just
reopen the subtree using the auto-open depth setting.  That's fairly
likely to be exactly what we want, and it's much easier.

This required a new CloseContents() API on IDiskImage and Partition
that closes the analyzed IFileSystem or IMultiPart object, without
closing the IChunkAccess.  This is the same state we'd be in after
AnalyzeDisk(..., ..., ChunksOnly).

All of these gymnastics are tied to the "enable writes" button.  If
the user just wants to look around, we don't need to do anything
fancy.

(The sector editor UI doesn't actually do much yet.  This just paves
the way for actual writing.)