Support copying of empty directories
Only works when copying to disk images, because we don't generally
deal with directory entries in file archives.

This fixes some inconsistent behavior in handling of Zip "directory"
files (which are currently defined as have a length of zero and
ending with a '/').  Also, new Zip records are now initialized with
the current date/time in the create and modification timestamp fields.