Implement "Close Sub-Tree"
This allows a disk image or file archive that has been opened and
made available in the archive tree to be closed.  Handy if something
was opened inadvertently.

Also, improved the "open best" function, which was descending into
the tree but not back-tracking.  This meant we'd start with the
selection on something useless in a multi-partition image with a
boring first partition (e.g. Opus ][).