Recursively extract/delete
We now handle extraction and deletion of directory trees.  In both
cases, disk images and archives that are open in the work tree will
be closed.

When viewing files, open archives are simply skipped over.  We now
show a message to that effect if the viewer can't find anything to
display.