Updates to file listing
Changed the way columns are configured so we can more accurately tune
the visible set to the thing we're looking at.  Added PathName
column, which is never shown at the same time as FileName.  The
full/dir toolbar buttons function but don't work quite right.

Minor tweaks to deletion: added special key handler so Del works when
a DataGrid has focus, and added a confirmation dialog.