Fill out directory tree and file list management
By default, for a hierarchical filesystem, the contents of the
selected directory are shown in the file list.  This adds a "full
list" mode that shows the contents of the entire filesystem.
The column configuration is switched from FileName to PathName.

We need to repopulate the directory tree and file list when contents
are added or removed.  We also need to update the file list when the
directory tree selection changes if we're in single-dir display
mode, and also when switching between single-dir and full-tree
mode.  This isn't free, so we do a quick check to see if the
currently-displayed contents match what we think they should be,
and repopulate the list if it doesn't match.

Selecting a different directory in the tree while in full-list mode
selects the appropriate directory in the file list.

Also, rename IFileEntry ChildCount to Count, and add a Count property
to IArchive.  In both cases they return the number of elements that
will be provided through their IEnumerable.