Implement file viewer "export" feature
Click the button to save the current contents of the viewer to a file.
If you're viewing a bitmap, you save a PNG.  If you're in hex dump
mode, you save a hex dump.

Also, reworked some of the simple-skipping gzip stuff in the CLI,
which was working when ExtArchive took care of it but wasn't working
right in "catalog".  The root of the problem is the absence of a
good filename inside the gzip archive itself, requiring us to pass
around the gzip archive's filename.

Also, correctly handle a zero-valued timestamp in gzip, which means
"no date".