Add no-name AppleSingle without ".as"
When adding AppleSingle files, we extract the contents, and use the
filename stored in the archive.  If the archive doesn't include a
filename, we use the name of the AppleSingle file itself.  However,
we don't want to keep the ".as" extension, since we're adding the
contents of the archive rather than the archive itself.