File viewer improvements
If a single file is double-clicked, select all available entries in
the file list (i.e. not directories or things that are open in the
work tree).  This allows the forward/backward buttons to work.

If the resource fork is being show, the Export button now saves that,
as "file.r.txt".