Added a "finishing" msg for add/delete progress
Also, spent some time trying to put the focus on the file list when
a file is first opened.  This turns out to be hard.

Putting the logical and/or keyboard focus on the DataGrid is easy,
but as soon as a key is hit the keyboard focus shifts to the
rightmost menu item.  So hitting the down arrow opens the Help or
DEBUG menu.  This behavior doesn't seem to change with different
approaches to setting the focus.  Removing the <Menu> element allows
focus to work as expected, but that's not really a solution.

Adding a button at the top of the DockPanel changed the behavior:
hitting the down arrow moved the focus to the button.  So somehow
the focus is always being placed on the first item in the DockPanel.
I tried using a Grid and defining the Menu last, but that didn't
change anything - focus goes to whatever is in row zero.

Some of the approaches indicate that it's necessary to put the focus
on a specific cell, rather than on the control or a row.  The code
wasn't able to find a row object during the various events I tried.
Adding a "jump" button to the toolbar did work: clicking on it set
the focus to the specified cell, and keyboard navigation worked
as expected.  So there's some chance that if we can figure out why
ItemContainerGenerator.ContainerFromIndex() is returning null we
can make this work.

Disabling virtualization and trying to UpdateLayout didn't help.

We may need an ItemContainerGenerator.StatusChanged hack.

If we want to revisit this issue, see:
 https://stackoverflow.com/q/29079570/294248
 https://stackoverflow.com/q/20537073/294248
 https://stackoverflow.com/q/3421597/294248
 https://stackoverflow.com/a/14551197/294248
 https://markjulmar.github.io/wpf/2008/09/04/part-2-changing-wpf-focus-in-code.html
 https://blog.magnusmontin.net/2013/11/08/how-to-programmatically-select-and-focus-a-row-or-cell-in-a-datagrid-in-wpf/
 GotKeyboardFocus / LostKeyboardFocus events