Implement "create directory" action
We need some text to explain the syntax.  Sticking it in the FS
implementation makes some sense (it's tied to the implementation),
but not entirely (UI should be provided by the app).  The right
answer is probably to define a set of characteristics that are
defined by the FS and then converted to words by the app, but I
don't think we need to get that fancy here.