File list DataGrid focus hack-a-thon
We want to focus on the file list when we open a new work file.
Also, if we double-click on a file list entry that holds a disk
image or file archive, we want to return to the file list after we
switch our archive tree selection to the item.

I tried unsuccessfully to make this work a couple of changes back.
This time I used the ItemContainerGenerator trick, which required
some extra hackishness because the focus change wouldn't stick
if it happened as the result of a ScrollIntoView call.  It needs
to happen as part of framework layout, so I don't clear the flag
unless I see UIElement.Measure in the stack trace.  (Ugh.)

This doesn't work when the DataGrid has no items, e.g. you deleted
the last file.  Normally this wouldn't matter, but the DataGrid's
context menu (right-click) doesn't work correctly if nothing has
focus.  (All items with "can execute" handlers are disabled; it
doesn't even try to run them.  Non-Command items work.)  This is
now hacked around with a ContextMenuOpening event handler that
detects the situation and puts the focus back on the file list.

This is kind of horrible to look at, but the application feels
much better.  This will likely come back to bite me.

Yay WPF.

Also, fixed CreateDirectory TextBox initial setup (wrong event).