Sketch out file entry attribute editor
Added some place holders.  There are two commands, one for the file
list, one for the directory list.  The directory list version is
how you change a filesystem's volume name.  (It could also be used
to edit a file archive header comment.)

Also, revised the definition of the storage size of a nonexistent
file part to be zero.

Also, tweaked the file entry selection stuff to not explode when you
pick a single directory and try to view it with right-click.

Also, tweaked some command enablers to check whether the object
they're about to modify is read-only.