Make scrolling to a directory entry work right
We use virtualization on the TreeViews used to show archives and
directories because WPF is pretty slow at creating them.  On large
disk images (e.g. A2Romulan) we can stall for an uncomfortable
amount of time while creating the UI elements.

The downside of virtualization is that you can't programmatically
select a TreeViewItem that hasn't been created yet, so jumping to
an entry in the tree when a directory is double-clicked in the file
list doesn't work if the destination is off-screen.

Fixing this required a few hours of research and experimentation,
not helped by Microsoft's removal of the sample code that everybody
references.  This seems to be working for the directory tree; it
needs to be extended to work for the archive tree too.

Yay WPF.