Progress on file attribute editing, part 5
Added HFS type editing.

Spent a little time playing with file list selection and focus.  If
you edit the filename of an HFS file, it can move in the list, but
you'd like it to still be selected and on-screen.  Similarly, after
deleting one or more files you want the selection to appear on the
item before the first deleted item, and after creating a directory
you want to focus on the new directory.

These now seem to work, but it's hard to keep the focus on the right
tree/list.  Double-clicking on a directory in the file list shifts
the focus to the directory tree, which may or may not be what is
desired.  It all feels a bit fragile.