Fiddle with formatting
This unifies some of the formatting for access flags, dates, and the
sizes of disk files.

Disk file sizes used to be shown compactly, e.g. "800KB", but many
recent UIs show it with a space ("800 KB") since that's a little
easier to read when you have a column of numbers.  For the "catalog"
and "mdc" output I think it looks better without the space.