Fix shrinking archive tree window
The left panel is split in two, with the archive tree on the top
and the directory tree on the bottom.  The settings file tracks
the sub-panel height, restoring it when the application is started.

This is harder than it sounds, because you can't just set the
sub-panel height to an absolute value.  The panels are part of a
Grid, with a splitter, and you have to use a proportional size
value or things get weird.  The previous approach was handling
it as proportional to the total height of the left panel, but WPF
apparently resizes the panel at some late stage, so the value you
read after the window has drawn is not the value you wrote.

The consequence is that the top sub-panel got a little bit smaller
every time the settings file was updated.

The approach that seems to work is recording the ratio of the sizes
of the top and bottom sub-panels, and restoring that.  It seems to
yield stable results.  Yay WPF.