Implement "add files" menu item, mostly
The standard Windows file dialogs have never allowed you to select
a collection of files and directories.  The original CiderPress
worked around this limitation by walking private data structures.
The correct solution here is to create a new dialog that works
correctly, but I'm not there yet.

For now, we allow files to be added but not directories.  To add
the contents of a directory, drag & drop is available.

Also, put a minimum size on unadorned sector images, to reduce some
of the noise when doing a bulk scan.