Support async cancel for file archive add ops
We need to check the cancel flag way down in the part source, so the
best way to handle this is to throw an exception.  We already have all
the necessary cleanup code for cases where the add operation fails,
so we just need to define an exception and catch it at the top.