Finish "new disk" dialog
Seems to be working.

Added "is this okay" tests to the disk image formats.

Fixed some exception handling and numeric overflows.  This was
mostly a matter of better exception handling in places that weren't
expected to be catching exceptions.

Gave the "is filename / volume name valid" functions a consistent
interface.