Finish "save as disk" command
This allows a disk image or partition to be saved as a new disk image,
with a type chosen by the user.  We copy sectors if possible, blocks
if not.  Failed reads result in zeroed-out data in the destination,
but don't halt the transfer.

Also, clarified some stuff in DiskArc.