Work around sub-volume sector edit issue
Some partitions can't be re-analyzed in isolation.  For example, a
DOS/ProDOS hybrid disk is essentially two filesystems that fully
overlap, but happen to mark each other's regions as in-use.  To
find both, we need to start at the top; re-analyzing the "inner"
filesystem might find the "outer" filesystem if we don't provide
context.

When sector-editing an embedded partition, we just close off that
partition.  We don't want to close the whole embedded zone because
that would lose the chunk subset that we're trying to edit.  The
correct approach to handling the update is to close the full set of
embedded volumes after the edit completes, and re-scan the set of
sub-volumes.

This is sort of annoying to implement, so I'm using a workaround: we
record the filesystem that was found initially, and only look for
that when we do the reanalysis.  This will usually work, because
replacing the filesystem (or damaging it beyond recogition) would
likely cause the embedded volume scanner to drop it anyway.