Add buttons for all disk actions in center panel
The various actions, like editing sectors and extracting disks, are
available from the Actions menu, from the right-click menu in the
archive tree window, and now as buttons in the center info panel.
The latter is the most convenient, and it's nice to have the options
laid out in front of you.  (Unless you're looking at a WOZ, there's
not much else going on in that screen anyway.)