Expand metadata interface
Added "description" and "value syntax" properties to the definition
of metadata items.  These are now displayed as tooltips when you
hover over entries in the GUI.

Added TestMetaValue() to the interface, so the GUI can test the
outcome of trying to set the value without actually changing it.

Also, 2IMG allowed deletion of the comment and version number entries
as a way to blank/unset them, but the definition of delete is "remove
the key from the set", which they weren't doing.  This is no longer
allowed.  The explanation of the version number field now explicitly
mentions -1, and the comment field can be erased by setting it to "".