Prototype UI for import/export conv opts editor
This will allow editing of the options that are passed to the import
and export converters.  The GUI should be the same as what is shown
by the file viewer.  These settings will affect file import/export,
whether performed by menu action or drag & drop.

All of the converters now have a verbose description string, which
we display in the GUI.