Fix scattered file deletion issue
Attempting to remove "subdir" and "subdir:file" would fail, because
"file" would be added to the set a second time when the filesystem
was traversed.  With duplicates excluded, it would also fail if the
display sort order was descending.  These problems weren't possible
in the "single directory" view, because it wasn't possible to select
both a directory and its contents.

We now screen out duplicates and sort the selection set when gathering
entries from a filesystem.  The behavior for file archives is
unchanged, because it's okay to delete directories without deleting
their contents.

Also, add a free space check at the end of the filesystem grinder,
to verify that we're not losing track of storage.