First cut at Trackstar disk images
Currently only works read-only.

Also, fix "one means all" file selection for file viewer, which got
mangled a bit by the previous change.  If we don't need directories
to come before their contents, then we don't need to sort the files.
If we're picking files for the file viewer, we probably want them
to appear in the selected order rather than a sort.

I did identify a bit of weird behavior where, if you select an
entry and then use Select All, WPF will put the initially-selected
entry at the start of the selection set.  This is a little weird
for the file viewer but hopefully it won't adversely affect anyone.