Be more cautious when reporting disk full
When adding files, if the disk runs out of space, we want to report
a "disk full" error to the user.  At the point we do this, the
filesystem is in a consistent state internally, but not everything
will have been flushed to the host file.  This creates a situation
where we potentially have a host file in an inconsistent state for
an extended period of time, increasing the risk that a crash or power
failure could leave the disk image in a bad state.

In particular: if you run out of space while adding files, and kill
the cp2_wpf process while the error dialog is shown, the directory
entry will be removed but the volume bitmap will still show the
blocks as being in use.  The issue appears to be that the disk image
Stream is buffering the last write call, so the volume bitmap change
hasn't made it to disk.  Disabling Stream buffering would fix the
issue, but with a general performance penalty.

We now defer reporting the error until after we have a chance to
flush changes to the disk image.