Implement physical disk access
This required some P/Invoke calls and figuring out how to do
privilege escalation.

It's not thoroughly tested -- my 2020s-era PC doesn't have the
variety of peripherals that my 2000s-era PC did -- but I was able
to copy files on and off of a CFFA-formatted CF card in a USB reader.

Side note: classic CiderPress is taking nearly a minute to open a
240MB CF card, while CP II does it in a fraction of a second.
It looks like some of the creaky old ScanCapacity code is taking
forever to run, but it's not even needed there.