Implement CP/M filesystem, part 3
5.25" CP/M disks use a sector skewing that is different from DOS or
ProDOS/Pascal.  The allocation unit on a CP/M disk is at least 1KB,
and we want to support large volumes with linear blocks, so we want
a block read function that reads at least 512 bytes.

This adds CP/M-order 512-byte block reads to the chunk access classes,
and updates the sector-edit features of the tools to be able to
specify CP/M blocks.

Also, fix ProDOS block reads on physically-ordered (.img) files.