Add Pascal and CP/M to cp2 tests
The "add file" and "copy file" tests are performed on all possible
filesystems.  Fixed a couple of bugs that were revealed.

Added a couple of entries to the create-disk-image test.

Added a CP/M disk image with every valid user number (0-15).

Also, rearranged "edit settings" in the GUI to allow setting
individual bits on ProDOS, NuFX, and CP/M files.

Also, in the GUI file viewer, keep track of the "raw DOS" setting.