Implement CP/M filesystem, part 9 (of 9)
Added support for handling user numbers as filename suffixes (e.g.
"FOO.TXT,3" for a text file in user area 3).  This was the most
natural approach, since user numbers are effectively an extension
of the filename.  It has some drawbacks, but it's straightforward
to implement and required minimal changes.

I briefly played around with virtual directories, which seem more
convenient for the user but are very intrusive and require lots of
CP/M-specific changes to application-level code.