Update sector editor
Rearranged the controls a bit, and added a combo box that lets you
select the sector skew.  This is useful for examining RDOS disks,
because they use 256-byte sectors but with ProDOS or physical skewing
rather than DOS.

We now show the sector codec on the screen, but I didn't actually
make it configurable (requires some library work).