Implement MacBinary, part 2 (of 2)
Implemented file detection, parsing, and file read.

Generally added "*.bin" and "*.macbin" as recognized extensions, but
we don't want to definitively declare ".bin" to mean MacBinary
because there are too many other things that use that extension in a
generic "this is a binary file" way.  Instead, we just let the
unknown-extension code dig through it.  This is less than ideal,
because the initial MacBinary format definition can false-positive
very easily, especially if you try to take into account files
generated by non-compliant implementations.

Added a freeware software disk as the example, because DiskCopy
images wrapped in MacBinary are pretty common.