Make Ctrl+C a little safer
Interrupting an operation that modifies a disk image is risky, for
the same reason that turning off the computer's power mid-write is
bad.  We can make this safer by flushing all pending writes before
querying the user (e.g. before overwriting an existing file), since
that's the most likely time for a kill signal to land.

This is a continuation of 1c86e112, which handled the specific case
of the GUI being killed while a "disk full" error dialog was being
displayed.  This change is more focused on the CLI, which is at
greater risk because the user can hit Ctrl+C in the middle of an add
or delete operation.  We use a .NET console feature to trap Ctrl+C
and just set a "cancel" flag that the progress dialog checks for.

Not all operations use the progress dialog, notably "mdc", so we leave
Ctrl+C set to the default behavior (terminate immediately) for
operations other than add/import/copy/move/delete.