Clean up filesystem content size reporting
The sizes of files on some filesystems are naturally described in
256-byte sectors, others in 512-byte blocks or just plain KiB.
This adds the appropriate logic for the recently-added filesystems.

Also, correctly report free space on MFS disks.

Also, remove the two sample programs (AddFile and ListContents)
from the distribution set.  These are source code samples, and don't
do anything that the "cp2" command doesn't.