Change "make bootable" to "reserve boot"
We can create disk images with five different filesystem types, and
only DOS is really made bootable.  The others are either always
as bootable as we can make them, e.g. ProDOS will always have a
boot block written, or we reserve the area without actually writing
an OS image, e.g. CP/M.

The GUI was already referring to it this way.  This changes the CLI
to match, and updates some variable names in the GUI innards.