Improve truncated archive checks
Added file length tests in the ArcReadStream uncompressed file
constructor and in the read routine at the point where EOF is hit.
For some types of compressed data the uncompressed length simply
isn't known; nothing we can do about those.

This is mostly of benefit as a secondary check for file archives with
uncompressed data.  If we're careless about checking lengths in the
file entry code, we might quietly interpret the EOF of a truncated
archive as the record EOF.  If the file entry implementation is
working correctly, this should never fire.

Also, rearranged the Dispose handling, which was complaining
unfairly when the GC cleaned up a partially-constructed object.