Experiment with drag & drop
The awkward one is multi-file drag from the file list, because the
DataGrid wants to un-select everything except the item you clicked
on.  We can hack around it by capturing the selection when the mouse
is clicked on a selected item, and then restoring it when the drag
starts.

Also, fix the magically expanding left panel.