Add "Copy" feature to file viewer
When the button is clicked or Ctrl+C is hit, the contents of the
file viewer are copied to the clipboard.  Plain text is copied as a
UnicodeText string, RTF is copied as both RTF and UnicodeText, CSV
is copied as CSV and UnicodeText, and images are copied as a
BitmapSource.  These may morph into other things under the hood,
e.g. the bitmap turns into three different things.

Excel seems to prefer the plain text to CSV, so when they're both
present it's necessary to use "paste special".  However, most things
(including Google Sheets running in Chrome) have no idea what to do
with CSV on the clipboard, so if you want to paste it into a text
file it's important to have both.

The data is copied from the current tab, so this works for resource
forks and notes.

Also, added a Find button and tweaked the GUI a little.