Try adding a note about DOS random text
There's no real differentiator between sequential- and random-access
text files under DOS 3.3.  It's slightly more complicated because
the file structure requires random-access text to be accessed in
"raw" mode.  It's not always obvious why things aren't working.

This adds a Note entry when random-access text conversion is attempted
on a DOS file without "raw" mode selected.

It might make sense to try to auto-select it when the Random Text
conversion is specified.  There are situations where "raw" mode is
desirable in other situations, e.g. looking at a 'B' file that has
data beyond the stated length, but I don't think it's ever useful
to view a random-access text file in non-raw mode.

It should be noted that random-access text files are rare, and that
there's no reliable way to recover the record length from DOS, so
we can't Just Make It Work.  But, maybe this tweak will give somebody
a hint as to why things aren't working for them.