Add "find" to file viewer, part 1
This is a first cut at implementing "find" for a RichTextBox.  This
is much harder than it sounds.  The current version does a simple
find previous/next string, wrapping around at the ends, with a
case-insensitive comparison.  It does not find strings that span
"runs", i.e. have different formatting.