Add "find" to file viewer, part 2 (of 2)
Added plain text search.  Much easier than rich text.  Still just
doing a simple case-insensitive string search.

Added F3 and Shift+F3 as find-next and find-prev keys.  When Ctrl+F
is hit, selection moves to the search string box.  Hitting Return in
the search string box also performs find-next, but because we have
to shift the focus to the text control it only does it once.  (We
could also try to map Return to find-next in the text control.)