Implement Print Shop clip art file converter
Converts clip art files from Print Shop and Print Shop GS.  Both
color and monochrome are handled.  The implementation is similar to
that in the original CiderPress, but an option has been added to
pixel-multiply the output by x2/x3 to match how the graphics look
when printed.

Added a few test files, found on disk images online.

Also, played with the bitmap scaling options, but ultimately didn't
make a change.  For some reason the "nearest neighbor" scaling is
behaving weirdly at 1:1, and it becomes very obvious with the dither
patterns in some of the clip art.  A screen grab showed that the
on-screen bitmap is displayed with the correct size.