Minor disk/fs tweaks
(1) Make it easier to override the DOS VTOC location.  A pair of
options in the AppHook can override the default T17 S0 with an
arbitrary value.  Useful for copy-protected disks that relocated
the the catalog.  (I didn't add anything to the application that
sets these options.)

(2) Added "patched" 13- and 16-sector codec entries.  These ignore
address/data epilogs and address header checksums, so a fair number
of copy-protected disks can be read directly.

(3) I found a ProDOS disk image that uses storage type 4 (ProFile
Pascal Manager storage).  We now show an appropriate message in Notes
when found.  Added a Notes entry for PPM.