Progress on drag & drop / copy & paste
Set up basic infrastructure.  We now augment the list of virtual
file streams with a set of serialized file attribute information.
The serialization uses JSON so we can display the contents easily
in the "drop target" debug dialog.

Also, added an IsValid property for IArchive instances, which didn't
have a way to report that they had been disposed of.