Implement file export via drag/copy
This requires more work up front than "extract" because we have to
set up the filename extension, which means we need to know what sort
of conversion we're going to be doing.  If the converter decides to
change its mind, we'll end up with output files that have the wrong
extension.

The good news is that we don't have to do the actual conversion
until it's time to drop/paste.  The bad news is we might have to do
the file extraction twice, which could be expensive for a large
compressed file.

Extract vs. export is now selected with a pair of toolbar buttons.
This is only for drag/copy; the extract/export actions continue to
do one specific thing.