Re-work clipboard formats
The original plan was to prepare the files for pasting into Windows
Explorer.  If the recipient was CiderPress II, it would have to
undo the preservation stuff to restore the original.  This had a
couple of problems, notably the potential loss of file attributes
and resource forks, but it turns out that it doesn't fit very well
with the existing "add files" code, which expects everything to live
on disk.

It turns out it's actually pretty easy to add a second set of virtual
file streams once you have all the code in place.  We now generate
a "foreign" file set for Windows apps, and an "xfer" file set that
is just for direct transfers.  The "xfer" set comes with serialized
metadata that includes the full set of file attributes.

We potentially create double the number of IStream references on the
clipboard, but I can't find anything that says there's a limit
(other than memory).  AFAIK they're not actually streams until
something tries to open them.