Improve drop/paste onto filesystems
Drag & drop onto filesystems now works.  It's similar to paste, but
the IDataObject can be used from multiple threads.

Copy & paste of a file onto itself is now blocked.  (It used to
delete the existing file and then try to open it for reading, which
is a pretty bad experience.)  This is not an issue for drag & drop,
since that just moves files between directories.

Dropping onto a directory in the file list now puts the files in
that directory, instead of always using the directory tree selection.