Add DOS text conversion
This is an optional conversion step that only applies when doing
a direct file transfer between filesystems (copy & paste or drag &
drop).  We set or clear the high bit on every byte in the text file.
An application setting can be used to enable/disable it.

It doesn't do anything when copying to a file archive because
we can't be sure what the files are supposed to look like in
the archive.  In practice it's probably safe to assume that all
archives store text files as plain ASCII, but rigging it up through
AddFileSource and its variants is a lot of work for a feature of
uncertain value.