Implement Apple Preferred Format SHR converter
APF images can have an arbitrary size and contain vendor-specific
data.  There's some ambiguity in the specification, and some
examples of non-compliant images, but I think this handles most of
the situations that arise.

The code handles MAIN and MULTIPAL.  I couldn't find an example of
MASK to test against, so that's not implemented.  It also handles
NOTE by copying the text to the Notes panel.