Add display of "host images" to the GUI
This allows GIF, JPEG, and PNG files to be displayed directly in the
file viewer.

The "export" button is disabled because making it work requires
manipulating the raw pixel data in the Windows media objects.  This
is doable but it's unclear whether it would be useful, since it's
better to extract rather than export.  We could make "export" do an
extract instead, but that might be confusing.

The "copy" button is also disabled because it's currently tied to the
export implementation.  We should be able to copy the Windows media
bitmap object to the clipboard.