Implement hi-res screen font converter
This converts an Apple II or Apple /// screen font to a bitmap with
the glyphs for all characters.  The Apple /// defined the high bit
as a way to control whether inverse mode was inverted or flashing,
so we optionally allow that to be displayed in a different color.

Added sample fonts from an Apple /// font disk, Apple Pascal, and
the DOS Toolkit.