Fiddle with copy & paste
Added "copy" to the context menu that appears when right-clicking
in the file list.  Rearranged the list so that the operations that
affect the selected files are separated from the operations that
don't (e.g. "create directory" doesn't care about the selection).

Having "paste" put the files into the directory that is selected in
the file list feels natural in many situations.  However, if the
root directory is filled with nothing but directories, it won't be
possible to paste there.  I'm not sure what the right answer is, so
I'm sticking with the current policy of always pasting into whatever
is selected in the directory tree.  For this reason I'm also not
adding "paste" to the file list context menu, but I am adding it to
the directory list context menu.