Add global "find file" feature
This allows files to be found across every open disk image and file
archive.  The search is by case-insensitive filename, and moves
forward or backward.

There are three minor annoyances.

First, the find operation needs a starting point, so if you don't
have an archive or filesystem open with an entry selected, it won't
work.  This means you can't start a find if you have an empty
archive open.

Second, the find order doesn't track the file list sort order.  If
you don't have things displayed in the order in which they appear
when read from the archive, the find result will appear to jump
around.

Third, the WPF selection change doesn't move the keyboard focus,
so the keyboard position will be out of sync with the selection.
(DataGrid focus has been an ongoing source of pain. Yay WPF.)

These issues are fixable, but would probably be best to handle as
part of a grander design rather than patching up find-next/prev.